go.crypto/ssh: add support for remote tcpip forwarding
Add support for server (remote) forwarded tcpip channels.
See RFC4254 Section 7.1

R=gustav.paul, jeff, agl, lieqiewang
CC=golang-dev
https://golang.org/cl/6038047

+++ ssh/channel.go
@@ -122,7 +122,7 @@ func (c *channel) Reject(reason RejectionReason, message string) error {
 
 	reject := channelOpenFailureMsg{
 		PeersId:  c.theirId,
-		Reason:   uint32(reason),
+		Reason:   reason,
 		Message:  message,
 		Language: "en",
 	}
